requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
outcome==1.3.0.post0
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-decouple==3.8
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
undetected-chromedriver==3.5.4
urllib3==2.1.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
